kconfig qconf: Add a proper link handling in ConfigInfoView
This creates a new class named ConfigInfoWindow
(based on ConfigSearchWindow but without the search bar)

ConfigInfoWindow is a window that displays the given config with its
help message and that allows to toggle the config (as in the search
window).

This window is used when someone click on a link in a ConfigInfoView
either from the main window or from another ConfigInfoWindow.

This commit also add a new slot in ConfigInfoView to handle both normal and debug links
and create/update the ConfigInfoWindow related to this ConfigInfoView.

Signed-off-by: Maxime Chretien <maxime.chretien@bootlin.com>