kconfig confdata: Add an option to keep all .config backups
When KCONFIG_KEEPALLBACKUPS is set in the environement,
instead of renaming the old .config to .config.old, kconfig put it in
a folder named .config.olds that stores all backups
with the name .config.oldX where X is a number.

The latest backup is the one with the highest number.

This is usefull to avoid doing a lot of manual backups to make sure to have
a working build when you make a lot of changes that can break everything.

Signed-off-by: Maxime Chretien <maxime.chretien@bootlin.com>